requirements.txt
chalice==1.26.5
pytest==6.2.5
aws-cdk-lib>=2.10.0
constructs>=10.0.0,<11.0.0